Rumi v2.0.1 release
Changes in this release:
- Support for arbitrary balancing area and balancing time for the input
  EC and the output DEC of an ECT (supply)
- Deterministic order of data in output csvs
- Bugfix: Emissions outputs exclude some STs
- Updated overview and demand documentation
- Bumped version to v2.0.1

In addition, the previous commit (v2.0.0) includes the following
changes to the supply and postprocessing modules that were not
acknowledged in that commit text. These changes continue to be supported
in this version (v2.0.1).
- Changes to address the following differences between Pyomo 5.x and Pyomo 6.x
- Support for changes to EndUseDemandEnergy parameter
- Support for user-defined constraints

For completeness, the following features mentioned in v2.0.0 commit text are
also supported in the supply and postprocess modules
- Support for non-EC based emissions
- Support for energy density and emissions of physical ECs varying by year